User sign in + registration works
When a user signs in, the postgreSQL database adds that user to its entries (if it's the user's first sign in and this user has a Lunatech domain).

Results showed: Upon multiple sign-ins amd sign-outs with XXXX@lunatech.nl, it's entry was made and did not dissappear or change unwantedly. Upon multiple attempts at signing in with XXXX@gmail.com, no entry was created and this user was denied access to the Lunatech Library system.